+++ hudai/__init__.py
@@ -8,11 +8,11 @@ from hudai.client import HudAi
 client = HudAi()
 """
 
-__version__ = '1.5.0'
-
-class HudAiError(Exception):
-    def __init__(self, message=None, type='validation_error'):
-        super(HudAiError, self).__init__(message)
-
-        self._message = message
-        self.type = type
+__package_name__ = 'hudai'
+__description__ = ('HUD.ai API python bindings')
+__keywords__ = ('hud.ai hudai foundry.ai foundrydc')
+__url__ = 'https://github.com/FoundryAI/hud-ai-python'
+__version__ = '2.0.0'
+__author__ = 'HUD.ai Engineering'
+__email__ = 'engineering@hud.ai'
+__license__ = 'MIT'
