--- hudai/__pkg__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Define project metadata
-"""
-
-__package_name__ = 'hudai'
-__description__ = ('HUD.ai API python bindings')
-__keywords__ = ('hud.ai hudai foundry.ai foundrydc')
-__url__ = 'https://github.com/FoundryAI/hud-ai-python'
-__version__ = '2.0.0'
-__author__ = 'HUD.ai Engineering'
-__email__ = 'engineering@hud.ai'
-__license__ = 'MIT'
