+++ hudai/resources/__init__.py
@@ -1,13 +1,13 @@
-from hudai.resources.article_highlights import ArticleHighlightsResource
-from hudai.resources.article_key_term import ArticleKeyTermResource
-from hudai.resources.article import ArticleResource
-from hudai.resources.company_key_term import CompanyKeyTermResource
-from hudai.resources.company import CompanyResource
-from hudai.resources.domain import DomainResource
-from hudai.resources.key_term import KeyTermResource
-from hudai.resources.system_event import SystemEventResource
-from hudai.resources.system_task import SystemTaskResource
-from hudai.resources.text_corpus import TextCorpusResource
-from hudai.resources.user_company import UserCompanyResource
-from hudai.resources.user_key_term import UserKeyTermResource
-from hudai.resources.user import UserResource
+from .article_highlights import ArticleHighlightsResource
+from .article_key_term import ArticleKeyTermResource
+from .article import ArticleResource
+from .company_key_term import CompanyKeyTermResource
+from .company import CompanyResource
+from .domain import DomainResource
+from .key_term import KeyTermResource
+from .system_event import SystemEventResource
+from .system_task import SystemTaskResource
+from .text_corpus import TextCorpusResource
+from .user_company import UserCompanyResource
+from .user_key_term import UserKeyTermResource
+from .user import UserResource
