+++ hudai/resources/system_event.py
@@ -1,4 +1,4 @@
-from hudai.resource import Resource
+from ..resource import Resource
 
 
 class SystemEventResource(Resource):
