+++ hudai/resources/system_task.py
@@ -1,4 +1,4 @@
-from hudai.resource import Resource
+from ..resource import Resource
 
 
 class SystemTaskResource(Resource):
