+++ hudai/resources/text_corpus.py
@@ -1,4 +1,4 @@
-from hudai.resource import Resource
+from ..resource import Resource
 
 
 class TextCorpusResource(Resource):
