+++ hudai/resources/user.py
@@ -1,4 +1,4 @@
-from hudai.resource import Resource
+from ..resource import Resource
 
 
 class UserResource(Resource):
