+++ hudai/resources/user_company.py
@@ -1,4 +1,4 @@
-from hudai.resource import Resource
+from ..resource import Resource
 
 
 class UserCompanyResource(Resource):
